refactor(layout): make all component svg icon and resturucture layout
